+++ 05_cub3D/libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: junghwki <junghwki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/03 16:38:17 by junghwki          #+#    #+#             */
-/*   Updated: 2020/10/13 20:46:59 by junghwki         ###   ########.fr       */
+/*   Updated: 2021/03/11 14:11:19 by junghwki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "../cub3d.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	index;
+	int	index;
 
 	index = 0;
 	while (s[index])
